Added time slow and dialed in basic air loco
E slows time so that i can see how the animations look in more detail. I set the blend times very low for jumping and landing, and start those animations around 40 percent into the anim to account for the windups and stuff in the animations. It still doesn't look perfect though. I might add a delay to pressing jump so i can play some of the wind up in the animation