Animation Continuation / Character Speed
Made the character's sprint speed slightly faster. Also going to continue with the animations. I made a blend space for idle/running jump but because I am setting the idle jump start position I can't use it in the blend. I am going to redo the graph to make it either/or. Idle jump or run jump no in between