GroundLocomotion Copied
The blueprint has been done over with the new skeleton. The abp has the whole event graph copied but everything else has to be done by hand. Also scaling the skeleton screwed up some of the modifications I made to the animations in blender so I have to fix that also. I hate Unreal with a burning passion