Camera tracks ragdoll
Wrote a bunch of code to reparent the spring arm but turns out I could just reparent it in the blueprint without messing anything up. So I stopped running the function in the blueprint but the code is still in the class. I'll probably just delete it when I need to actually use scripting for something