Running parent constructor in Another Go Character
In an attempt to fix the bp forgetting some settings. I added the Initializer Super() that runs the Character class constructor. Maybe that does good stuff that I am not doing in my inherited class. Also this is a stopping point before I attempt to parent the camera spring arm to a specific bone in the skeletal mesh. Right now I can parent it in the editor to either the capsule collider (cam wont follow ragdoll) or the skeletal mesh (parents to root and gets stuck in geometry) I am going to try to parent it to the head bone of the skeletal mesh in AnotherGoCharacter constructor