Nothing works nvm
Can't seem to parent to the bones. Unless I brute force it and copy the bone position every frame somehow? The problem is the fucking root bone. Turning off collision on the camera fixes the problem so It must be that it is getting forced into geometry. And it is parented to the root so that has to be what is putting it in stuff